requirements.txt
# Core scientific
numpy>=1.24
pandas>=2.0
scipy>=1.10
PyWavelets>=1.4

# Deep learning
torch>=2.0

# EEG processing
mne>=1.5

# Classical ML
scikit-learn>=1.3
lightgbm>=4.0
xgboost>=2.0

# TabNet
pytorch-tabnet>=4.1

# Hyperparameter tuning
optuna>=3.0

# Acceleration
numba>=0.57
joblib>=1.3

# Plotting
matplotlib>=3.7

# Config
PyYAML>=6.0

# Utilities
tqdm>=4.65
